+++ test/unit/CQWebsocket.test.js
@@ -4,7 +4,7 @@ const { WebsocketState } = CQWebSocket
 const { test } = require('ava')
 
 test('new Websocket() with default options', function (t) {
-  t.plan(7)
+  t.plan(8)
 
   const bot = new CQWebSocket()
 
@@ -21,10 +21,11 @@ test('new Websocket() with default options', function (t) {
   t.deepEqual(bot._wsOptions, {
     fragmentOutgoingMessages: false
   })
+  t.deepEqual(bot._requestOptions, { })
 })
 
 test('new Websocket() with custom options', function (t) {
-  t.plan(7)
+  t.plan(8)
 
   const bot = new CQWebSocket({
     enableAPI: false,
@@ -36,7 +37,10 @@ test('new Websocket() with custom options', function (t) {
     reconnectionAttempts: 10,
     reconnectionDelay: 5000,
     fragmentOutgoingMessages: true,
-    fragmentationThreshold: 0x4000
+    fragmentationThreshold: 0x4000,
+    requestOptions: {
+      timeout: 2000
+    }
   })
 
   t.is(bot._monitor.EVENT.state, WebsocketState.INIT)
@@ -44,6 +48,9 @@ test('new Websocket() with custom options', function (t) {
   t.is(bot._baseUrl, '8.8.8.8:8888/ws')
   t.is(bot._qq, 123456789)
   t.is(bot._token, 'qwerasdf')
+  t.deepEqual(bot._requestOptions, {
+    timeout: 2000
+  })
   t.deepEqual(bot._reconnectOptions, {
     reconnection: true,
     reconnectionAttempts: 10,