+++ test/unit/call.test.js
@@ -11,6 +11,7 @@ test.cb('#__call__(method, params)', function (t) {
   const preSpy = spy()
   const postSpy = spy()
   const apiResponseSpy = spy()
+  // provide qq to avoid invoking `_apiSock.send`
   const bot = new CQWebsocket({ qq: 123456789 })
     .on('api.send.pre', preSpy)
     .on('api.send.post', postSpy)
@@ -83,6 +84,7 @@ test('#__call__() while disconnected.', async function (t) {
 test.cb('#__call__(method, params, options) with timeout option', function (t) {
   t.plan(5)
 
+  // provide qq to avoid invoking `_apiSock.send`
   const bot = new CQWebsocket({ qq: 123456789 })
     .on('ready', function () {
       t.is(bot._responseHandlers.size, 0)
@@ -105,6 +107,26 @@ test.cb('#__call__(method, params, options) with timeout option', function (t) {
         })
         // response handler should be reset due to timeout
         t.is(bot._responseHandlers.size, 0)
+        t.end()
+      })
+    })
+    .connect()
+})
+
+test.cb('#__call__(method) use global request options if options is omitted', function (t) {
+  t.plan(2)
+
+  // provide qq to avoid invoking `_apiSock.send`
+  const bot = new CQWebsocket({ qq: 123456789, requestOptions: { timeout: 2000 } })
+
+  let start
+  bot
+    .on('ready', function () {
+      start = Date.now()
+      bot('test').catch(err => {
+        t.true(err instanceof ApiTimoutError)
+        // Not sure if this assertion is stable (?)
+        t.is(Math.round((Date.now() - start) / 1000), 2)
         t.end()
       })
     })